Fix #1589
